moved NumberLiteral
