+++ src_smartcontract/sc/CodeElement.cpp
@@ -57,14 +57,13 @@
 #include "sc_expression/ConstructorArray.h"
 #include "sc_expression/ConstructorCall.h"
 #include "sc_expression/LiteralExpression.h"
-#include "sc_expression/NumberLiteral.h"
 #include "sc_expression/BooleanLiteral.h"
 #include "sc_expression/ParenthesisExpression.h"
 #include "sc_expression/VariableIdentifier.h"
 #include "sc_expression/FunctionCallExpression.h"
 
 #include "sc_expression_literal/NullLiteral.h"
-
+#include "sc_expression_literal/NumberLiteral.h"
 
 #include "sc_expression_bit/AndExpression.h"
 #include "sc_expression_bit/BitReverseExpression.h"
