+++ src_smartcontract/sc_expression/CMakeLists.txt
@@ -12,7 +12,6 @@ set(__src
 	FunctionCallExpression.cpp
 	LiteralExpression.cpp
 	MemberReferenceExpression.cpp
-	NumberLiteral.cpp
 	ParenthesisExpression.cpp
 	VariableIdentifier.cpp
 )
