+++ src_smartcontract/sc_expression_literal/CMakeLists.txt
@@ -2,6 +2,7 @@
 
 set(__src
 	NullLiteral.cpp
+	NumberLiteral.cpp
 )
 
 handle_sub(codablecashlib "${__src}" smartcontract sc_expression_literal)
